refactor: extract VpatCriteriaRow from vpat-criteria-table
